[NEWS][BACK] Dodanie cache'a dla modułu News

+++ ejb-web/src/main/java/com/stobinski/bottlecaps/ejb/rest/NewsController.java
@@ -43,7 +43,7 @@ public class NewsController {
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("count")
-	public Long getNewsCount() {
+	public int getNewsCount() {
 		return newsManager.countNews();
 	}
